+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+* improved `SparqlSourceProcessor`: add rudimentary rate limit handling
+
 ## [3.1.2] - 2025-08-26
 
 ### Fixed